+++ client/src/pages/Board.js
@@ -1,0 +1,7 @@
+function Board() {
+  return (  
+    <div>Board</div>
+  );
+}
+
+export default Board;
